MAINT: backend: remove pyproject_toml
